+++ src/test/java/christmas/domain/event/EventTest.java
@@ -2,18 +2,6 @@ package christmas.domain.event;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import christmas.controller.MenuCount;
-import christmas.domain.Date;
-import christmas.domain.Menus;
-import christmas.domain.Order;
-import christmas.domain.menu.Menu;
-import java.util.Map;
-import java.util.Optional;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-
 class EventTest {
 
 }
